fix: handle empty object as input
3 verifiers trigger an error when an empty object is passed down:
- OpenAttestationEthereumDocumentStoreStatus
- OpenAttestationDnsTxt
- OpenAttestationEthereumTokenRegistryStatus
For all those verifiers the error happens during the test function (which determines if a verifier should run or not).
The plan is to fix those test functions so that the verifiers doesn't run if issuers property (in that case) is not provided => the result for those verifiers will be skipped fragment.
On top of that OpenAttestationHash verifier currently return invalid while it shouldn't run because targethash/merkleroot/data are missing => the plan is to fix this behaviour so that the verifier is skipped in that case.
Also to make sure that verifiers test function handle correctly, we should provide a partial object (deeply)

closes #111